requirements.txt
customtkinter
Pillow
soundfile
sounddevice
numpy
